rename the origin jobmaster to jobmaster demo
